--- a/build.gradle
+++ b/build.gradle
@@ -7,18 +7,17 @@ buildscript {
     dependencies {
         classpath 'com.github.jengelman.gradle.plugins:shadow:1.2.3'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.6'
-        classpath 'com.netflix.nebula:gradle-dependency-lock-plugin:4.4.0'
-        classpath 'com.netflix.nebula:nebula-dependency-recommender:4.0.2'
+        classpath 'com.netflix.nebula:gradle-dependency-lock-plugin:4.9.5'
+        classpath 'com.netflix.nebula:nebula-dependency-recommender:5.0.0'
         classpath 'com.netflix.nebula:nebula-publishing-plugin:4.9.1'
-        classpath 'com.palantir.baseline:gradle-baseline-java:0.14.0'
+        classpath 'com.palantir.baseline:gradle-baseline-java:0.17.1'
         classpath 'gradle.plugin.com.palantir.configurationresolver:gradle-configuration-resolver-plugin:0.2.0'
-        classpath 'net.ltgt.gradle:gradle-errorprone-plugin:0.0.9'
     }
 }
 
 plugins {
-    id 'com.palantir.git-version' version '0.2.0'
-    id 'org.inferred.processors' version '1.2.2'
+    id 'com.palantir.git-version' version '0.9.1'
+    id 'org.inferred.processors' version '1.2.15'
 }
 
 apply plugin: 'com.palantir.baseline-config' // provides baselineUpdateConfig task
@@ -29,7 +28,7 @@ repositories {
 }
 
 dependencies {
-    baseline "com.palantir.baseline:gradle-baseline-java-config:0.14.0@zip"
+    baseline "com.palantir.baseline:gradle-baseline-java-config:0.17.1@zip"
 }
 
 allprojects {
@@ -49,7 +48,7 @@ subprojects {
     apply plugin: 'com.palantir.baseline-checkstyle'
     apply plugin: 'com.palantir.baseline-eclipse'
     apply plugin: 'com.palantir.baseline-idea'
-    apply plugin: 'net.ltgt.errorprone'
+    apply plugin: 'com.palantir.baseline-error-prone'
 
     repositories {
         jcenter()