--- a/build.gradle
+++ b/build.gradle
@@ -63,15 +63,18 @@ subprojects {
         lockFile = 'versions.lock'
         configurationNames = ['runtime', 'compileClasspath']
     }
-    task verifyDependencyLocksAreCurrent << {
-        def expectedDependencies = tasks.saveLock.getOutputLock()
-        def actualDependencies = tasks.saveLock.getGeneratedLock()
-        def digester = java.security.MessageDigest.getInstance('SHA')
-        logger.info("Verifying integrity of dependency locks: {} vs {}", expectedDependencies, actualDependencies)
-        if (digester.digest(expectedDependencies.bytes) != digester.digest(actualDependencies.bytes)) {
-            throw new GradleException("The dependencies of project " + project.name + " do not match the expected "
-                + "dependencies recorded in " + expectedDependencies + ". "
-                + "Run `./gradlew generateLock saveLock` and commit the updated version.lock files")
+
+    task verifyDependencyLocksAreCurrent {
+        doLast {
+            def expectedDependencies = tasks.saveLock.getOutputLock()
+            def actualDependencies = tasks.saveLock.getGeneratedLock()
+            def digester = java.security.MessageDigest.getInstance('SHA')
+            logger.info("Verifying integrity of dependency locks: {} vs {}", expectedDependencies, actualDependencies)
+            if (digester.digest(expectedDependencies.bytes) != digester.digest(actualDependencies.bytes)) {
+                throw new GradleException("The dependencies of project " + project.name + " do not match the expected "
+                    + "dependencies recorded in " + expectedDependencies + ". "
+                    + "Run `./gradlew generateLock saveLock` and commit the updated version.lock files")
+            }
         }
     }