--- a/extras/jackson-support/src/main/java/com/palantir/remoting3/ext/jackson/ShimJdk7Module.java
+++ b/extras/jackson-support/src/main/java/com/palantir/remoting3/ext/jackson/ShimJdk7Module.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Palantir Technologies, Inc. All rights reserved.
+ * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.