--- a/http2-agent/src/main/java/com/palantir/remoting3/http2/Http2Agent.java
+++ b/http2-agent/src/main/java/com/palantir/remoting3/http2/Http2Agent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Palantir Technologies, Inc. All rights reserved.
+ * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
  */
 
 package com.palantir.remoting3.http2;