--- a/jaxrs-clients/src/main/java/com/palantir/remoting3/jaxrs/feignimpl/PathTemplateHeaderEnrichmentContract.java
+++ b/jaxrs-clients/src/main/java/com/palantir/remoting3/jaxrs/feignimpl/PathTemplateHeaderEnrichmentContract.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
  */
 
 package com.palantir.remoting3.jaxrs.feignimpl;