--- a/jaxrs-clients/src/main/java/feign/CborDelegateDecoder.java
+++ b/jaxrs-clients/src/main/java/feign/CborDelegateDecoder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Palantir Technologies, Inc. All rights reserved.
+ * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
  */
 
 package feign;