--- a/jersey-servers/src/main/java/com/palantir/remoting3/servers/jersey/GuavaOptionalMessageBodyWriter.java
+++ b/jersey-servers/src/main/java/com/palantir/remoting3/servers/jersey/GuavaOptionalMessageBodyWriter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Palantir Technologies, Inc. All rights reserved.
+ * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
  */
 
 /*