--- a/retrofit2-clients/src/main/java/com/palantir/remoting3/retrofit2/CborConverterFactory.java
+++ b/retrofit2-clients/src/main/java/com/palantir/remoting3/retrofit2/CborConverterFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Palantir Technologies, Inc. All rights reserved.
+ * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
  */
 
 package com.palantir.remoting3.retrofit2;