… .baseline/checkstyle/checkstyle-suppressions.xml
… .baseline/checkstyle/checkstyle.xml
… .baseline/copyright/001_apache-2.0.txt
… .baseline/copyright/999_palantir.txt
… .baseline/eclipse/org.eclipse.jdt.core.prefs
… .baseline/eclipse/static/dotfile.settings/org.eclipse.jdt.ui.prefs
… .baseline/idea/intellij-java-palantir-style.xml
… circle.yml
… error-handling/src/test/java/com/palantir/remoting3/errors/SerializableErrorErrorDecoderTests.java
… extras/jackson-support/src/test/java/com/palantir/remoting3/ext/jackson/ObjectMappersTest.java
… extras/jackson-support/src/test/java/com/palantir/remoting3/ext/jackson/ShimJdk7ModuleTest.java
… extras/refresh-utils/src/test/java/com/palantir/remoting3/ext/refresh/RefreshableProxyInvocationHandlerTest.java
… extras/refresh-utils/src/test/java/com/palantir/remoting3/ext/refresh/RefreshableTest.java
… gradle/wrapper/gradle-wrapper.jar
… http-clients/src/test/java/com/palantir/remoting3/clients/ClientConfigurationsTest.java
… http-clients/src/test/java/com/palantir/remoting3/clients/UserAgentTest.java
… http2-agent/src/test/java/com/palantir/remoting3/http2/Http2AgentTests.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/JaxRsClientConfigRefreshTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/JaxRsClientFailoverTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/JaxRsClientGuavaOptionalHandlingTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/JaxRsClientJava8OptionalHandlingTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/JaxRsClientMeshProxyConfigTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/JaxRsClientProxyConfigTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/TestBase.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/TestService.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/TracerTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/feignimpl/GuavaOptionalAwareDecoderTest.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/feignimpl/GuavaOptionalComplexType.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/feignimpl/GuavaTestServer.java
… jaxrs-clients/src/test/java/com/palantir/remoting3/jaxrs/feignimpl/HeaderAccessUtilsTest.java