I just tried to run `./gradlew idea` locally and it didn't complete within 8 minutes.  Hoping that gradle's new parallelization work will help! (Also CircleCI builds are now 1 minute faster)

## Before: ./gradlew idea is unbearably slow

```
./gradlew idea --parallel --profile
Parallel execution is an incubating feature.

> Configure project : 
The Task.leftShift(Closure) method has been deprecated and is scheduled to be removed in Gradle 5.0. Please use Task.doLast(Action) instead.
        at build_3r25vddp4g5pl6xrvibv6w5a4$_run_closure4.doCall(/Users/dfox/http-remoting/build.gradle:66)
        (Run with --stacktrace to get the full stack trace of this deprecation warning.)

> Configure project :keystores 
The ConfigurableReport.setDestination(Object) method has been deprecated and is scheduled to be removed in Gradle 5.0. Please use the method ConfigurableReport.setDestination(File) instead.
        at build_cu65bkiqxvshne6ouloo5dw6e.configureTestTask(/Users/dfox/http-remoting/keystores/build.gradle:44)
        (Run with --stacktrace to get the full stack trace of this deprecation warning.)
<========-----> 63% EXECUTING [2m 51s]
<========-----> 63% EXECUTING [2m 51s]
<========-----> 63% EXECUTING [8m 31s]
> :okhttp-clients:ideaModule > Resolve dependencies :okhttp-clients:runtimeClasspath
> :okhttp-clients:ideaModule > Resolve dependencies :okhttp-clients:runtimeClasspath
> :error-handling:ideaModule Resolve files of :http-clients:compileClasspathntimeClasspath
> :http-clients:ideaModule > Resolve files of :http-clients:compileClasspathntimeClasspath
> :retrofit2-clients:ideaModule > Resolve dependencies :retrofit2-clients:runtimeClasspathpath
> :jaxrs-clients:ideaModule > Resolve dependencies :jaxrs-clients:runtimeClasspathimeClasspath
> :jaxrs-scala-clients:ideaModule > Resolve dependencies :jaxrs-scala-clients:runtimeClasspath
> :jersey-servers:ideaModule > Resolve dependencies :jersey-servers:runtimeClasspath
^C%  
```

## After: ./gradlew idea takes < 1 second:

```
git clean -xdf && time ./gradlew idea --parallel --rerun-tasks
Removing .gradle/
Removing error-handling/error-handling.iml
Removing extras/extras.iml
Removing extras/jackson-support/jackson-support.iml
Removing extras/refresh-utils/refresh-utils.iml
Removing http-clients/http-clients.iml
Removing http-remoting.iml
Removing http-remoting.ipr
Removing http-remoting.iws
Removing http2-agent/http2-agent.iml
Removing jaxrs-clients/jaxrs-clients.iml
Removing jaxrs-scala-clients/jaxrs-scala-clients.iml
Removing jersey-servers/jersey-servers.iml
Removing keystores/keystores.iml
Removing okhttp-clients/okhttp-clients.iml
Removing pkcs1-reader-bouncy-castle/pkcs1-reader-bouncy-castle.iml
Removing pkcs1-reader-sun/pkcs1-reader-sun.iml
Removing retrofit2-clients/retrofit2-clients.iml
Removing tracing-jaxrs/tracing-jaxrs.iml
Removing tracing-okhttp3/tracing-okhttp3.iml
Removing tracing/tracing.iml
Parallel execution is an incubating feature.

BUILD SUCCESSFUL in 0s
20 actionable tasks: 20 executed
./gradlew idea --parallel --rerun-tasks  1.17s user 0.14s system 90% cpu 1.455 total
```

@uschi2000